pyproject.toml
[project]
name = "employee-tasks"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django-extensions (>=4.1,<5.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
black = "^25.9.0"
mypy = "^1.18.2"
flake8 = "^7.3.0"
isort = "^6.0.1"


[tool.poetry.group.test.dependencies]
coverage = "^7.10.7"

